Feat: Allow client to fetch procedures data
